+++ src/main/java/com/google/idea/perf/TracerConfig.kt
@@ -17,6 +17,7 @@
 package com.google.idea.perf
 
 import com.google.idea.perf.agent.MethodListener
+import com.google.idea.perf.util.ConcurrentAppendOnlyList
 import org.objectweb.asm.Type
 import java.lang.reflect.Method
 import java.util.concurrent.locks.ReentrantLock
