+++ src/main/java/com/google/idea/perf/util/Util.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.util
 
 import java.text.NumberFormat
